Refactor user model references to use custom user model throughout the codebase
