+++ changelog.d/20260810_100000_aleksei_sosov_custom_user_model.md
@@ -1,0 +1,5 @@
+### Changed
+
+- CVAT now uses its own user model instead of Django's default one.
+  Existing users, permissions, and data are unchanged
+  (<https://github.com/cvat-ai/cvat/pull/11023>)
