+++ cvat/apps/dataset_manager/tests/test_streaming_efficiency.py
@@ -10,7 +10,7 @@ from datumaro import AnnotationType, Bbox, LabelCategories
 from datumaro.components import media
 from datumaro.components.dataset import StreamDataset
 from datumaro.components.dataset_base import CategoriesInfo, DatasetBase, DatasetItem
-from django.contrib.auth.models import Group, User
+from django.contrib.auth.models import Group
 from rest_framework import status
 from rq.job import Job as RQJob
 
@@ -35,6 +35,7 @@ from cvat.apps.engine.tests.utils import (
     generate_image_file,
     get_paginated_collection,
 )
+from cvat.apps.iam.models import User
 
 
 class TestExtractors(TestCase):
