+++ cvat/apps/engine/tests/test_rest_api_3D.py
@@ -14,7 +14,7 @@ from collections import defaultdict
 from glob import glob
 from shutil import copyfile
 
-from django.contrib.auth.models import Group, User
+from django.contrib.auth.models import Group
 from rest_framework import status
 
 from cvat.apps.dataset_manager.task import TaskAnnotation
@@ -27,6 +27,7 @@ from cvat.apps.engine.tests.utils import (
     ImportApiTestBase,
     get_paginated_collection,
 )
+from cvat.apps.iam.models import User
 
 from .utils import check_annotation_response
