+++ cvat/apps/engine/user_deletion.py
@@ -8,7 +8,8 @@ from typing import TypedDict
 from django.db import transaction
 from django.db.models import Count, Q, QuerySet
 
-from cvat.apps.engine.models import CloudStorage, Project, Task, User
+from cvat.apps.engine.models import CloudStorage, Project, Task
+from cvat.apps.iam.models import User
 from cvat.apps.organizations.models import Organization
 from cvat.utils.django_database.utils import _ModelT
