+++ cvat/apps/growth/models.py
@@ -2,13 +2,13 @@
 #
 # SPDX-License-Identifier: MIT
 
-from django.contrib.auth.models import User
+from django.conf import settings
 from django.db import models
 
 
 class UserGrowthData(models.Model):
     user = models.OneToOneField(
-        User,
+        settings.AUTH_USER_MODEL,
         on_delete=models.CASCADE,
         related_name="growth_data",
     )
