+++ cvat/apps/growth/signals.py
@@ -2,9 +2,10 @@
 #
 # SPDX-License-Identifier: MIT
 
-from django.contrib.auth.models import User
 from django.db.models.signals import post_save
 from django.dispatch import receiver
+
+from cvat.apps.iam.models import User
 
 from .models import UserGrowthData
