+++ cvat/apps/iam/admin.py
@@ -5,10 +5,11 @@
 
 from django.contrib import admin
 from django.contrib.auth.admin import GroupAdmin, UserAdmin
-from django.contrib.auth.models import Group, User
+from django.contrib.auth.models import Group
 from django.utils.translation import gettext_lazy as _
 
 from cvat.apps.engine.models import Profile
+from cvat.apps.iam.models import User
 
 
 class ProfileInline(admin.StackedInline):
@@ -60,7 +61,6 @@ class CustomGroupAdmin(GroupAdmin):
     fieldsets = ((None, {"fields": ("name",)}),)
 
 
-admin.site.unregister(User)
 admin.site.unregister(Group)
 admin.site.register(User, CustomUserAdmin)
 admin.site.register(Group, CustomGroupAdmin)
